// src/shared/enum/conspiracy-sub-types.ts
export enum ConspiracySubTypes {
	SINGLE_USE = '<a href="/conspiracy-subtypes#Single_Use_Conspiracies">Single Use</a>',
	REUSABLE = '<a href="/conspiracy-subtypes#Reusable_Conspiracies">Reusable</a>',
	END_OF_DAY = '<a href="/conspiracy-subtypes#End_of_Day_Conspiracies">End of Day</a>',
	FRIEND = '<a href="/conspiracy-subtypes#Friend_or_Foe_Conspiracies">Friend</a>',
	FOE = '<a href="/conspiracy-subtypes#Friend_or_Foe_Conspiracies">Foe</a>',
}